refactor: update comments schema and related imports
- Changed import paths for comments schema across multiple files to ensure consistency and maintainability.
- Removed unused comment-related DTOs and schemas to streamline the codebase and eliminate redundancy.
