+++ src/articles/articles.controller.ts
@@ -1,5 +1,4 @@
 import { setupArticles } from '@articles/articles.module';
-import { CommentResponseDto, CreateCommentDto } from '@comments/dto';
 import { Elysia, t } from 'elysia';
 import {
   ArticleFeedQueryDto,
@@ -10,7 +9,6 @@ import {
   UpdateArticleDto,
 } from './dto';
 import {
-  toCommentResponse,
   toCreateArticleInput,
   toFeedResponse,
   toResponse,
@@ -174,66 +172,6 @@ export const articlesController = new Elysia().use(setupArticles).group(
                 tokenAuth: [],
               },
             ],
-          },
-        },
-      )
-      .post(
-        '/:slug/comments',
-        async ({ body, params, store, request }) => {
-          const comment = await store.commentsService.createComment(
-            params.slug,
-            body.comment,
-            await store.authService.getUserIdFromHeader(request.headers),
-          );
-          return toCommentResponse(comment);
-        },
-        {
-          beforeHandle: app.store.authService.requireLogin,
-          body: CreateCommentDto,
-          response: CommentResponseDto,
-          detail: {
-            summary: 'Add Comments to an Article',
-          },
-        },
-      )
-      .get(
-        '/:slug/comments',
-        async ({ params, store, request }) => {
-          const userId = await store.authService.getOptionalUserIdFromHeader(
-            request.headers,
-          );
-          const comments = await store.commentsService.getComments(
-            params.slug,
-            userId === null ? undefined : userId,
-          );
-          return { comments: comments.map(toCommentResponse) };
-        },
-        {
-          response: t.Object({
-            comments: t.Array(CommentResponseDto),
-          }),
-          detail: {
-            summary: 'Get Comments from an Article',
-          },
-        },
-      )
-      .delete(
-        '/:slug/comments/:id',
-        async ({ params, store, request }) => {
-          await store.commentsService.deleteComment(
-            params.slug,
-            Number.parseInt(params.id, 10),
-            await store.authService.getUserIdFromHeader(request.headers),
-          );
-        },
-        {
-          beforeHandle: app.store.authService.requireLogin,
-          params: t.Object({
-            slug: t.String(),
-            id: t.String(),
-          }),
-          detail: {
-            summary: 'Delete Comment',
           },
         },
       )
