+++ src/articles/articles.schema.ts
@@ -1,5 +1,5 @@
 import { users } from '@/users/users.model';
-import { comments } from '@comments/schema';
+import { comments } from '@comments/comments.schema';
 import { articleTags } from '@tags/tags.model';
 import { relations } from 'drizzle-orm';
 import {
