+++ src/comments/comments.repository.ts
@@ -1,8 +1,8 @@
 import type { Database } from '@/database.providers';
 import { articles } from '@articles/articles.schema';
 import type { ArticleRow } from '@articles/interfaces/article-row.interface';
-import { comments } from '@comments/schema/comments.schema';
 import { and, desc, eq } from 'drizzle-orm';
+import { comments } from './comments.schema';
 
 // TODO: Move & Re-evaluate this type. It's really just a band-aid.
 type CreateCommentDto = {
