--- src/comments/dto/comment-author.dto.ts
@@ -1,8 +1,0 @@
-import { Type } from '@sinclair/typebox';
-
-export const CommentAuthorDto = Type.Object({
-  username: Type.String(),
-  bio: Type.String(),
-  image: Type.String(),
-  following: Type.Boolean(),
-});
