+++ src/comments/dto/index.ts
@@ -1,3 +1,2 @@
 export * from './create-comment.dto';
 export * from './comment-response.dto';
-export * from './comment-author.dto';
