+++ src/comments/mappers/index.ts
@@ -1,0 +1,1 @@
+export * from './comments.mapper';
