--- src/comments/schema/index.ts
@@ -1,2 +1,0 @@
-export * from './comments.schema';
-export * from './comment-relations.schema';
