+++ src/database.providers.ts
@@ -1,5 +1,5 @@
 import * as articlesSchema from '@articles/articles.schema';
-import * as commentsSchema from '@comments/schema';
+import * as commentsSchema from '@comments/comments.schema';
 import { dbCredentialsString } from '@db/config';
 import * as tagsSchema from '@tags/tags.model';
 import * as usersSchema from '@users/users.model';
